Minor bug fixes and app logo added

+++ app/src/main/java/com/nerdytech/noteit/auth/LoginActivity.java
@@ -3,6 +3,8 @@ package com.nerdytech.noteit.auth;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
+import android.text.TextUtils;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
@@ -22,11 +24,12 @@ import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.nerdytech.noteit.MainActivity;
 import com.nerdytech.noteit.R;
+import com.nerdytech.noteit.Splash;
 
 public class LoginActivity extends AppCompatActivity {
     EditText lEmail,lPassword;
     Button loginNow;
-    TextView forgetPass,createAcc;
+    TextView forgetPass,createAcc,anon;
     FirebaseAuth fAuth;
     FirebaseFirestore fStore;
     FirebaseUser user;
@@ -48,13 +51,37 @@ public class LoginActivity extends AppCompatActivity {
 
         forgetPass = findViewById(R.id.forgotPasword);
         createAcc = findViewById(R.id.createAccount);
+        anon=findViewById(R.id.anonymous);
         user = FirebaseAuth.getInstance().getCurrentUser();
 
         fAuth = FirebaseAuth.getInstance();
         fStore = FirebaseFirestore.getInstance();
 
-
-        showWarning();
+        Intent intent=getIntent();
+        final int fromSplash=intent.getIntExtra("splash",0);
+        if(fromSplash==0) {
+            showWarning();
+        }
+        anon.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                // create new anonymous account
+                fAuth.signInAnonymously().addOnSuccessListener(new OnSuccessListener<AuthResult>() {
+                    @Override
+                    public void onSuccess(AuthResult authResult) {
+                        Toast.makeText(LoginActivity.this, "Logged in With Temporary Account.", Toast.LENGTH_SHORT).show();
+                        startActivity(new Intent(getApplicationContext(),MainActivity.class));
+                        finish();
+                    }
+                }).addOnFailureListener(new OnFailureListener() {
+                    @Override
+                    public void onFailure(@NonNull Exception e) {
+                        Toast.makeText(LoginActivity.this, "Error ! " + e.getMessage(), Toast.LENGTH_SHORT).show();
+                        finish();
+                    }
+                });
+            }
+        });
 
         loginNow.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -62,7 +89,7 @@ public class LoginActivity extends AppCompatActivity {
                 String mEmail = lEmail.getText().toString();
                 String mPassword = lPassword.getText().toString();
 
-                if(mEmail.isEmpty() || mPassword.isEmpty()){
+                if (mEmail.isEmpty() || mPassword.isEmpty()) {
                     Toast.makeText(LoginActivity.this, "Fields Are Required.", Toast.LENGTH_SHORT).show();
                     return;
                 }
@@ -71,25 +98,35 @@ public class LoginActivity extends AppCompatActivity {
 
                 spinner.setVisibility(View.VISIBLE);
 
-                if(fAuth.getCurrentUser().isAnonymous()){
-                    FirebaseUser user = fAuth.getCurrentUser();
+                if (fromSplash == 0){
+                    if (fAuth.getCurrentUser().isAnonymous()) {
+                        final FirebaseUser user = fAuth.getCurrentUser();
 
-                    fStore.collection("notes").document(user.getUid()).delete().addOnSuccessListener(new OnSuccessListener<Void>() {
-                        @Override
-                        public void onSuccess(Void aVoid) {
-                            Toast.makeText(LoginActivity.this, "All Temp Notes are Deleted.", Toast.LENGTH_SHORT).show();
-                        }
-                    });
+                        fStore.collection("notes").document(user.getUid()).delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+                            @Override
+                            public void onSuccess(Void aVoid) {
+                                Toast.makeText(LoginActivity.this, "All Temp Notes are Deleted.", Toast.LENGTH_LONG).show();
+                                // delete Temp user
 
-                    // delete Temp user
+                                user.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+                                    @Override
+                                    public void onSuccess(Void aVoid) {
+                                        Toast.makeText(LoginActivity.this, "Temp user Deleted.", Toast.LENGTH_SHORT).show();
+                                    }
+                                });
+                            }
+                        }).addOnFailureListener(new OnFailureListener() {
+                            @Override
+                            public void onFailure(@NonNull Exception e) {
+                                Toast.makeText(LoginActivity.this, "Error in Deleting Note.", Toast.LENGTH_SHORT).show();
+                            }
+                        });;
+                        //delete temp uesr data
 
-                    user.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
-                        @Override
-                        public void onSuccess(Void aVoid) {
-                            Toast.makeText(LoginActivity.this, "Temp user Deleted.", Toast.LENGTH_SHORT).show();
-                        }
-                    });
-                }
+
+
+                    }
+            }
 
                 fAuth.signInWithEmailAndPassword(mEmail,mPassword).addOnSuccessListener(new OnSuccessListener<AuthResult>() {
                     @Override
@@ -134,5 +171,13 @@ public class LoginActivity extends AppCompatActivity {
                 });
 
         warning.show();
+    }
+    @Override
+    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
+        if(item.getItemId() == android.R.id.home){
+            Toast.makeText(this, "Good Bye!", Toast.LENGTH_SHORT).show();
+            onBackPressed();
+        }
+        return super.onOptionsItemSelected(item);
     }
 }
